♻️ refactor: toolbar.tsx と style-panel.tsx をディレクトリ構成に分割
- toolbar.tsx (729行) → toolbar/ (6ファイル: toolbar, tool-button, export-menu, status-bar, copilot-toggle, voice-button)
- style-panel.tsx (448行) → style-panel/ (3ファイル: style-panel, compact-color-picker, index)

AIが各コンポーネントを独立して読み書きできるサイズに分割。
