+++ apps/web/src/components/style-panel/index.ts
@@ -1,0 +1,1 @@
+export { StylePanel } from "./style-panel.js";
